Updated SearchBestPath View and its helper functions
Updated the Search Best Optimal Path and its helper functions. The changes is done to to add three optimal paths instead of just one. This is due to the fact that some of the target stations and hydrants may not be in good condition leaving the choice for the user to re-select other closest station to contact. The frontend was adjusted as such